requirements.txt
pip=22.1.2
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
python-telegram-bot-calendar=1.0.5
deep-translator=1.8.3
requests~=2.28.0
psycopg2=2.9.3